The flask passes arguments to the robot
The flask passes arguments on the setting page to the robot. This version is not final since variables of the flask app need to follow the 'name' input of HTML file. I will modify them after HTML pages complete. 
Since the flask needs to import ssh.py, ssh.py needs to be moved in the same directory.