Added a function to terminate auto-mode.py and run auto-mode.py in the background
I added a function to terminate auto-mode.py. we run it before we execute the auto-mode.py every time. It can avoid 2 scripts accessing motors at the same time.
I added '&' to the run_auto_mode function, which makes the auto-mode script run in the background and does not occupy the terminal, then we can run other scripts such as uv.py.